+++ db/migrate/20190207155855_add_number_to_sites.rb
@@ -1,0 +1,6 @@
+class AddNumberToSites < ActiveRecord::Migration[6.0]
+  def change
+    add_column :sites, :number, :integer
+    add_index :sites, :number
+  end
+end