+++ test/controllers/editor/sites_controller_test.rb
@@ -1,8 +1,9 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require "test_helper"
 
-class Editor::SitesControllerTest < ActionController::TestCase
+# Allows editors to manage sites.
+class Editor::SitesControllerTest < ActionDispatch::IntegrationTest
   setup do
     @editor = users(:editor)
     @viewer = users(:viewer)
@@ -10,119 +11,123 @@ class Editor::SitesControllerTest < ActionController::TestCase
   end
 
   def site_params
-    { name: 'Site Three', slug: 'site-three', address: @site.address }
+    {
+      name: "Site Three",
+      slug: "site-three",
+      address: "123 Road Way",
+      number: "3"
+    }
   end
 
-  test 'should get index as editor' do
+  test "should get index as editor" do
     login(@editor)
-    get :index
+    get editor_sites_url
     assert_response :success
     assert_not_nil assigns(:sites)
   end
 
-  test 'should not get index as viewer' do
+  test "should not get index as viewer" do
     login(@viewer)
-    get :index
-    assert_redirected_to dashboard_path
+    get editor_sites_url
+    assert_redirected_to dashboard_url
   end
 
-  test 'should get new as editor' do
+  test "should get new as editor" do
     login(@editor)
-    get :new
+    get new_editor_site_url
     assert_response :success
   end
 
-  test 'should not get new as viewer' do
+  test "should not get new as viewer" do
     login(@viewer)
-    get :new
-    assert_redirected_to dashboard_path
+    get new_editor_site_url
+    assert_redirected_to dashboard_url
   end
 
-  test 'should create site as editor' do
+  test "should create site as editor" do
     login(@editor)
-    assert_difference('Site.count') do
-      post :create, params: { site: site_params }
+    assert_difference("Site.count") do
+      post editor_sites_url, params: { site: site_params }
     end
-    assert_redirected_to editor_site_path(assigns(:site))
+    assert_redirected_to editor_site_url(assigns(:site))
   end
 
-  test 'should not create site with blank name' do
+  test "should not create site with blank name" do
     login(@editor)
-    assert_difference('Site.count', 0) do
-      post :create, params: { site: site_params.merge(name: '') }
+    assert_difference("Site.count", 0) do
+      post editor_sites_url, params: { site: site_params.merge(name: "") }
     end
-    assert_not_nil assigns(:site)
-    assert assigns(:site).errors.size > 0
-    assert_equal ["can't be blank"], assigns(:site).errors[:name]
-    assert_template 'new'
-  end
-
-  test 'should not create site as viewer' do
-    login(@viewer)
-    assert_difference('Site.count', 0) do
-      post :create, params: { site: site_params }
-    end
-    assert_redirected_to dashboard_path
-  end
-
-  test 'should show site as editor' do
-    login(@editor)
-    get :show, params: { id: @site }
     assert_response :success
   end
 
-  test 'should not show site as viewer' do
+  test "should not create site as viewer" do
     login(@viewer)
-    get :show, params: { id: @site }
-    assert_redirected_to dashboard_path
+    assert_difference("Site.count", 0) do
+      post editor_sites_url, params: { site: site_params }
+    end
+    assert_redirected_to dashboard_url
   end
 
-  test 'should get edit as editor' do
+  test "should show site as editor" do
     login(@editor)
-    get :edit, params: { id: @site }
+    get editor_site_url(@site)
     assert_response :success
   end
 
-  test 'should not get edit as viewer' do
+  test "should not show site as viewer" do
     login(@viewer)
-    get :edit, params: { id: @site }
-    assert_redirected_to dashboard_path
+    get editor_site_url(@site)
+    assert_redirected_to dashboard_url
   end
 
-  test 'should update site as editor' do
+  test "should get edit as editor" do
     login(@editor)
-    patch :update, params: { id: @site, site: site_params }
-    assert_redirected_to editor_site_path(assigns(:site))
+    get edit_editor_site_url(@site)
+    assert_response :success
   end
 
-  test 'should update site with blank name' do
-    login(@editor)
-    patch :update, params: { id: @site, site: site_params.merge(name: '') }
-    assert_not_nil assigns(:site)
-    assert assigns(:site).errors.size > 0
-    assert_equal ["can't be blank"], assigns(:site).errors[:name]
-    assert_template 'edit'
-  end
-
-  test 'should not update site as viewer' do
+  test "should not get edit as viewer" do
     login(@viewer)
-    patch :update, params: { id: @site, site: site_params }
-    assert_redirected_to dashboard_path
+    get edit_editor_site_url(@site)
+    assert_redirected_to dashboard_url
   end
 
-  test 'should destroy site as editor' do
+  test "should update site as editor" do
     login(@editor)
-    assert_difference('Site.current.count', -1) do
-      delete :destroy, params: { id: @site }
+    patch editor_site_url(@site), params: { site: site_params }
+    @site.reload
+    assert_equal "Site Three", @site.name
+    assert_equal "site-three", @site.slug
+    assert_equal "123 Road Way", @site.address
+    assert_equal 3, @site.number
+    assert_redirected_to editor_site_url(assigns(:site))
+  end
+
+  test "should update site with blank name" do
+    login(@editor)
+    patch editor_site_url(@site), params: { site: site_params.merge(name: "") }
+    assert_response :success
+  end
+
+  test "should not update site as viewer" do
+    login(@viewer)
+    patch editor_site_url(@site), params: { site: site_params }
+    assert_redirected_to dashboard_url
+  end
+
+  test "should destroy site as editor" do
+    login(@editor)
+    assert_difference("Site.current.count", -1) do
+      delete editor_site_url(@site)
     end
-    assert_redirected_to editor_sites_path
+    assert_redirected_to editor_sites_url
   end
 
-  test 'should not destroy site as viewer' do
+  test "should not destroy site as viewer" do
     login(@viewer)
-    assert_difference('Site.current.count', 0) do
-      delete :destroy, params: { id: @site }
+    assert_difference("Site.current.count", 0) do
+      delete editor_site_url(@site)
     end
-    assert_redirected_to dashboard_path
+    assert_redirected_to dashboard_url
   end
 end